# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Live Test",
        "description": "This is api testing",
        "contact": {
            "email": "xyz@xyz.com"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/stat": {
            "get": {
                "tags": [
                    "Statistic"
                ],
                "summary": "Get Statistics",
                "operationId": "get-statistics",
                "parameters": [
                    {
                        "name": "mytest",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
...
}